pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "1.0.0"
description = ""
authors = ["vlad moryakov <sat1sfy@vk.com>"]
packages = [{include = "gendiff"}]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^6.2.3"
pytest-cov = "^3.0.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
PyYAML = "^6.0"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gen_diff:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
